Deploy task no longer relies on private manifest file
It will make it possible for people without access private repository to
understand a duplicate this ci build since the private repository only
holds secrets and the current deploy state.

In order to persist the bosh-init state we need to move the state file
between the persistance repo(bosh-concourse-ci) and the manifest
directory

Since the manifest is now rendered from a HEREDOC directly in the deploy
task, we no longer need to move the releases and stemcells around to
match the value in the manifest.  Instead we can just have the manifest
reference the original locations of those assets.

[#97441782](https://www.pivotaltracker.com/story/show/97441782)

Signed-off-by: Megan Murawski <megan.hyland@emc.com>