Removes dependency on private bats config
We now generate the bats config file directly in the run-bats task to
enable people without access to our private repo to be able to
understand and run these tasks themselves.

Also removes param for setting the BAT_VCAP_PASSWORD as this is coupled
to the password injected into the director manifest, which is currently
hardcoded to be c1oudc0w so we could not change this value anyway.

[#97441782](https://www.pivotaltracker.com/story/show/97441782)

Signed-off-by: Megan Murawski <megan.hyland@emc.com>