Delete Vapp in delete_vm if vapp name == vm.agent_id
In create_vm, if vapp has a requested name, the vapp is permanent. If
not, the vapp is temporary and should be removed when the vm is deleted.
The vapp name will be the requested name if it is a requested vapp, else
it will be the agent_id of the vm it is created for.

Removed some tests that tested for old behavior, which was to delete all
vapps when the last vm running on that vapp was deleted.

[#100476064](https://www.pivotaltracker.com/story/show/100476064)

Signed-off-by: Xuebin He <xuebin.he@emc.com>