Handle old failed tasks on vCloud Entities
vCloud maintains a list of historical tasks for a vApp. These tasks are
returned in the Xml element of a GET vAPP response.
When BOSH recomposes a vApp (to merge in new VMs for example) currently
it aborts the BOSH Task if any failed tasks are identified.
This PR enables a configurable time threshold @old_task_threshold which
tells the vcd_client library to disregard tasks older than the value (in
minutes). By default tasks older than an hour will be ignored.

This is currently not configurable as it is not exposed in the
[spec file](jobs/vcloud_cpi/spec) and used in the
[cpi.json.erb](jobs/vcloud_cpi/templates/cpi.json.erb).

[#119488293](https://www.pivotaltracker.com/story/show/119488293)